feat: include .cdx.gz extension in ZipNum .idx
Changes the ZipNum format generation to include the .cdx.gz extension in
all file references.
So we don't need to use the .loc file on pywb configuration and opens
the possibility to use the prefix configuration that could just apply a
basic regex to read the cdx.gz files from a remote location.

Changes:
- flat_cdxj_to_zipnum.py: Modified .idx file generation to keep .cdx.gz
  extension in part column (shard name field) instead of stripping it
- flat_cdxj_to_zipnum.py: Modified .loc file generation to include .cdx.gz
  extension in both columns (previously only second column had it)
- Updated all documentation to reflect correct format:
  * .idx format: <key>\t<shard_name.cdx.gz>\t<offset>\t<length>\t<shard_num>
  * .loc format: <shard_name.cdx.gz>\t<filepath> (both columns identical)